# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CofradeNet API
  description: |
    API REST de la plataforma CofradeNet. Gestiona hermandades, bandas, procesiones,
    usuarios cofrades y el panel de administración de la plataforma.

    ## Autenticación
    La mayoría de los endpoints de escritura requieren un token JWT Bearer.
    Obtén el token con `POST /auth/login` e inclúyelo en la cabecera:
    `Authorization: Bearer <token>`

    ## Roles disponibles
    | Rol | Descripción |
    |-----|-------------|
    | `cofrade` | Usuario registrado. Puede seguir entidades, crear publicaciones y gestionar su perfil. |
    | `hermandad` | Administra el perfil de su hermandad y sus procesiones. |
    | `banda` | Administra el perfil de su banda y su agenda de actuaciones. |
    | `admin` | Acceso completo a la plataforma. |
  version: 1.0.0
  contact:
    name: Equipo CofradeNet
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /seguimientos/hermandad/{id}:
    get:
      tags: [seguimientos]
      summary: Estado de seguimiento de una hermandad
      description: Devuelve si el usuario autenticado sigue la hermandad y el total de seguidores.
      operationId: estadoSeguimientoHermandad
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Estado de seguimiento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeguimientoEstado'
    post:
      tags: [seguimientos]
      summary: Seguir hermandad
      operationId: seguirHermandad
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: Seguimiento registrado.
    delete:
      tags: [seguimientos]
      summary: Dejar de seguir hermandad
      operationId: dejarHermandad
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Seguimiento eliminado.
  /seguimientos/banda/{id}:
    get:
      tags: [seguimientos]
      summary: Estado de seguimiento de una banda
      operationId: estadoSeguimientoBanda
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Estado de seguimiento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeguimientoEstado'
    post:
      tags: [seguimientos]
      summary: Seguir banda
      operationId: seguirBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: Seguimiento registrado.
    delete:
      tags: [seguimientos]
      summary: Dejar de seguir banda
      operationId: dejarBanda
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Seguimiento eliminado.
  /seguimientos/usuario/{id}:
    get:
      tags: [seguimientos]
      summary: Estado de seguimiento de un usuario
      operationId: estadoSeguimientoUsuario
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Estado de seguimiento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeguimientoEstado'
    post:
      tags: [seguimientos]
      summary: Seguir usuario
      operationId: seguirUsuario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '201':
          description: Seguimiento registrado.
    delete:
      tags: [seguimientos]
      summary: Dejar de seguir usuario
      operationId: dejarUsuario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Seguimiento eliminado.

  # ── PUBLICACIONES ─────────────────────────
  /publicaciones:
    post:
      tags: [publicaciones]
      summary: Crear publicación
      operationId: crearPublicacion
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublicacionRequest'
      responses:
        '201':
          description: Publicación creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publicacion'
  /publicaciones/hermandad/{id}:
    get:
      tags: [publicaciones]
      summary: Publicaciones de una hermandad
      operationId: publicacionesPorHermandad
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de publicaciones ordenada por fecha descendente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publicacion'
  /publicaciones/banda/{id}:
    get:
      tags: [publicaciones]
      summary: Publicaciones de una banda
      operationId: publicacionesPorBanda
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de publicaciones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publicacion'
  /publicaciones/usuario/{id}:
    get:
      tags: [publicaciones]
      summary: Publicaciones de un usuario
      operationId: publicacionesPorUsuario
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de publicaciones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publicacion'
  /publicaciones/{id}:
    delete:
      tags: [publicaciones]
      summary: Eliminar publicación (autor o ADMIN)
      operationId: eliminarPublicacion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Publicación eliminada.
        '403':
          $ref: '#/components/responses/Error403'

  # ── SEARCH ────────────────────────────────
...
  /admin/publicaciones/{id}:
    delete:
      tags: [admin]
      summary: Eliminar cualquier publicación (moderación)
      operationId: adminEliminarPublicacion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Publicación eliminada.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en `POST /auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador numérico del recurso.
  responses:
...
    Error403:
      description: Sin permisos — el usuario no tiene el rol necesario.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    RolUsuario:
      type: string
      enum: [admin, cofrade, hermandad, banda]
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        rol:
          $ref: '#/components/schemas/RolUsuario'
        verificado:
          type: boolean
        ciudadResidencia:
          $ref: '#/components/schemas/Ciudad'
...
    Hermandad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        nombrePopular:
          type: string
        titulares:
          type: array
          items:
            type: string
        añoFundacion:
          type: integer
          nullable: true
        templo:
          type: string
        diaSalida:
          type: string
          nullable: true
        descripcion:
          type: string
          nullable: true
        imagenEscudo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'

    # ── Bandas ────────────────────────────────
...
    Banda:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        estiloMusical:
          type: string
        historia:
          type: string
          nullable: true
        numeroComponentes:
          type: integer
          nullable: true
        imagenLogo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'
        repertorio:
          type: array
          items:
            $ref: '#/components/schemas/Marcha'
        eventos:
          type: array
          items:
            $ref: '#/components/schemas/Evento'

    # ── Procesiones ───────────────────────────
...
    Evento:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        fechaHora:
          type: string
          format: date-time
        lugar:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
        bandaId:
          type: integer
...
    CreatePublicacionRequest:
      type: object
      required: [contenido]
      properties:
        contenido:
          type: string
          example: "Este año nuestra hermandad ha incorporado a la Banda de la Macarena."
        tipo:
          type: string
          enum: [general, itinerario]
          default: general
        hermandadId:
          type: integer
          nullable: true
        bandaId:
          type: integer
          nullable: true
    Publicacion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        contenido:
          type: string
        tipo:
          type: string
          enum: [general, itinerario]
        fechaCreacion:
          type: string
          format: date-time
          readOnly: true
        autor:
          $ref: '#/components/schemas/Usuario'
        hermandad:
          $ref: '#/components/schemas/Hermandad'
          nullable: true
        banda:
          $ref: '#/components/schemas/Banda'
          nullable: true

    # ── Seguimientos ──────────────────────────
    SeguimientoEstado:
      type: object
      properties:
        sigues:
          type: boolean
          description: Si el usuario autenticado sigue esta entidad.
        seguidores:
          type: integer
          description: Número total de seguidores.

    # ── Search ────────────────────────────────
...
    Ciudad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        provincia:
          type: string
        pais:
          type: string
...
    Marcha:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        compositor:
          type: string
          nullable: true

    # ── Admin ─────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "La procesión no existe"
        error:
          type: string
          example: "Not Found"
security:
  - BearerAuth: []